wip: move nativeenv.go back to it's own file
